refactor: extract MCP tool definitions into separate files
We need to do some refactoring of our MediaWiki API MCP. The goal is to extract the MCP tool function definitions from `server.py` in put them into separate files in a dedicated directory, e.g. `server_tools/wiki_page_get.py`, `server_tools/wiki_page_edit.py`, etc.

```git-revs
6adc51e  (Base revision)
8d534ca  Create server_tools package init file
407ae9b  Extract wiki_page_edit tool to separate file
9cd0b25  Extract wiki_page_get tool to separate file
52df2ce  Extract wiki_search tool to separate file
ffd31ef  Extract wiki_opensearch tool to separate file
44351d0  Extract wiki_page_move tool to separate file
6d7dbb5  Extract wiki_page_delete tool to separate file
32ea7a7  Extract wiki_page_undelete tool to separate file
6d8cbe9  Extract wiki_meta_siteinfo tool to separate file
572baa4  Refactor server.py to use extracted tool registration functions
98ba05c  Update server_tools __init__.py to export registration functions
534478a  Snapshot before auto-test
cf098de  Fix docstring indentation in wiki_search tool to match test expectations
a384561  Snapshot before codemcp change
17d566f  Fix imports and callable type annotation for wiki_page_edit
9050f59  Snapshot before codemcp change
f09e521  Fix imports and callable type annotation for wiki_page_get
945fef7  Snapshot before codemcp change
7d03f78  Fix imports and callable type annotation for wiki_search
df0c0ad  Snapshot before codemcp change
5d4fd63  Fix imports and callable type annotation for wiki_opensearch
cedf68f  Snapshot before codemcp change
7f0c1fd  Fix imports and callable type annotation for wiki_page_move
91486b4  Snapshot before codemcp change
4fff9a1  Fix imports and callable type annotation for wiki_page_delete
0986dfe  Snapshot before codemcp change
1f22d10  Fix imports and callable type annotation for wiki_page_undelete
696d55b  Snapshot before codemcp change
67335e9  Fix imports and callable type annotation for wiki_meta_siteinfo
HEAD     Snapshot before auto-test
```

codemcp-id: 33-refactor-extract-mcp-tool-definitions-into-separat
